+++ requirements.txt
@@ -1,9 +1,19 @@
+alembic==0.8.3
 Flask==0.10.1
 Flask-HTTPAuth==2.7.0
+flask-marshmallow==0.6.2
+Flask-Migrate==1.6.0
+Flask-Script==2.0.5
 Flask-SQLAlchemy==2.0
 itsdangerous==0.24
 Jinja2==2.8
+Mako==1.0.2
 MarkupSafe==0.23
+marshmallow==2.1.3
+marshmallow-sqlalchemy==0.6.0
 passlib==1.6.5
+psycopg2==2.6.1
+python-editor==0.4
+six==1.10.0
 SQLAlchemy==1.0.8
 Werkzeug==0.10.4
